fix: Changed the int64 value representing the warehouse ID to int32.

+++ internal/application/dto/request/create_item.go
@@ -6,7 +6,7 @@ import (
 )
 
 type CreateItemDTO struct {
-	WarehouseID int64        `json:"warehouse_id"`
+	WarehouseID int32        `json:"warehouse_id"`
 	Code        string       `json:"code"`
 	Name        string       `json:"name"`
 	Description string       `json:"description"`
